# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LandOverSEA - International Dating & Cultural Connection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /language/quizzes:
    get:
      operationId: getLanguageQuizzes
      tags: [language-learning]
      summary: List available quizzes
      parameters:
        - name: language
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
            description: vocabulary | grammar | comprehension | cultural
        - name: difficulty
          in: query
          schema:
            type: ["string", "null"]
            description: beginner | intermediate | advanced
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LanguageQuiz"
  /language/quizzes/{quizId}:
    get:
      operationId: getLanguageQuiz
      tags: [language-learning]
      summary: Get a quiz with its public questions
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguageQuiz"
        "404":
          description: Quiz not found
  /language/quizzes/{quizId}/attempt:
    post:
      operationId: submitQuizAttempt
      tags: [language-learning]
      summary: Submit a quiz attempt with answers
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptResult"
  /language/ai-coach:
    post:
      operationId: sendAiCoachMessage
      tags: [language-learning]
      summary: Send a message to the AI language coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCoachInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCoachResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 30
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/push-tokens:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register or reactivate a push-delivery token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "201":
          description: Push token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushToken"
    delete:
      operationId: removePushToken
      tags: [notifications]
      summary: Remove a push-delivery token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Token removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: Get server-side notification delivery preferences
      responses:
        "200":
          description: Current preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "400":
          description: Query parameters are not supported
    put:
      operationId: updateNotificationPreferences
      tags: [notifications]
      summary: Update server-side notification delivery preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferencesInput"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "400":
          description: Invalid or incomplete preferences
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── AUTH ──────────────────────────────────────────────────────────────
...
    LanguageQuiz:
      type: object
      required: [id, title, language, category, questionCount, difficulty, xpReward, questions]
      properties:
        id:
          type: number
        title:
          type: string
        language:
          type: string
        category:
          type: string
          enum: [vocabulary, grammar, comprehension, cultural]
        questionCount:
          type: number
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        xpReward:
          type: number
        questions:
          type: array
          items:
            $ref: "#/components/schemas/LanguageQuizQuestion"
    LanguageQuizQuestion:
      type: object
      required: [id, text, options]
      properties:
        id:
          type: number
        text:
          type: string
        options:
          type: array
          items:
            type: string
    QuizAttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, answer]
            properties:
              questionId:
                type: number
              answer:
                type: string
    QuizAttemptResult:
      type: object
      required: [score, total, passed, xpEarned]
      properties:
        score:
          type: number
        total:
          type: number
        passed:
          type: boolean
        xpEarned:
          type: number
        feedback:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: number
              correct:
                type: boolean
              correctAnswer:
                type: string
    AiCoachInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        targetLanguage:
          type: ["string", "null"]
        mode:
          type: string
          enum: [practice, correction, cultural, freeform]
          default: freeform
    AiCoachResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        corrections:
          type: array
          items:
            type: string
        vocabulary:
          type: array
          items:
            type: string
        culturalNote:
          type: ["string", "null"]
...
    Notification:
      type: object
      required: [id, userId, type, isRead, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        type:
          type: string
          enum: [match, message, like, superlike, verification, premium, booking, system]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        relatedId:
          type: ["number", "null"]
        relatedType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PushTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
        platform:
          type: string
        provider:
          type: string
    PushToken:
      type: object
      required: [id, platform, provider, isActive]
      properties:
        id:
          type: number
        platform:
          type: string
        provider:
          type: string
        isActive:
          type: boolean
    NotificationPreferencesInput:
      type: object
      additionalProperties: false
      required: [pushEnabled, emailEnabled, bookingEnabled]
      properties:
        pushEnabled:
          type: boolean
        emailEnabled:
          type: boolean
        bookingEnabled:
          type: boolean
        showOnlineStatus:
          type: boolean
        readReceipts:
          type: boolean
    NotificationPreferences:
      type: object
      additionalProperties: false
      required: [pushEnabled, emailEnabled, bookingEnabled, showOnlineStatus, readReceipts]
      properties:
        pushEnabled:
          type: boolean
        emailEnabled:
          type: boolean
        bookingEnabled:
          type: boolean
        showOnlineStatus:
          type: boolean
        readReceipts:
          type: boolean
...